+++ src/templates/tsconfig.json.ts
@@ -49,8 +49,8 @@ export = (answers => {
 		"**/*.d.ts"
 	],
 	"exclude": [
-		${useTypeScript ? `"build/**",` : ""}
-		"node_modules/**",
+		${useTypeScript ? (`"build/**",
+		`) : ""}"node_modules/**",
 		"admin/**"
 	]
 }`;